Moved files from inside scene folder to one folder above. Continuing with building the renderer WIP
